make tests work 2

+++ src/controllers/post.ts
@@ -1,6 +1,6 @@
 import { BaseController } from "./base_controller";
 import { Response } from "express";
-import { AuthResquest } from "../common/auth_middleware";
+import { AuthRequest } from "../common/auth_middleware";
 import Post, { IPost } from "../models/post";
 
 class PostController extends BaseController<IPost>{
@@ -8,23 +8,21 @@ class PostController extends BaseController<IPost>{
         super(Post)
     }
 
-    async post(req: AuthResquest, res: Response) {
+    async post(req: AuthRequest, res: Response) {
         const _id = req.user._id;
-        req.body.owner = _id;
+        req.body.createdBy = _id;
         super.post(req, res);
     }
 
-    async get(req: AuthResquest, res: Response) {
+    async get(req: AuthRequest, res: Response) {
         super.get(req, res);
     }
 
-    async getById(req: AuthResquest, res: Response) {
-        const _id = req.user._id;
-        req.body.owner = _id;
+    async getById(req: AuthRequest, res: Response) {
         super.getById(req, res);
     }
 
-    async getByUserId(req: AuthResquest, res: Response) {
+    async getByUserId(req: AuthRequest, res: Response) {
         try {
             const posts = await this.model.find({ createdBy: req.params.user_id });
             res.send(posts);
@@ -33,28 +31,40 @@ class PostController extends BaseController<IPost>{
         }
     }
 
-    async putById(req: AuthResquest, res: Response) {
-        const _id = req.user._id;
-        req.body.owner = _id;
-        super.putById(req, res);
+    async putById(req: AuthRequest, res: Response) {
+        try {
+            const post = await Post.findById(req.body._id).select('createdBy');
+            if (post.createdBy.toString() === req.user._id) await super.putById(req, res);
+            else {
+                res.status(401).send();
+            }
+        } catch (err) {
+            res.status(400).send(err.message);
+        }
     }
 
-    async deleteById(req: AuthResquest, res: Response) {
-        const _id = req.user._id;
-        req.body.owner = _id;
-        super.deleteById(req, res);
+    async deleteById(req: AuthRequest, res: Response) {
+        try {
+            const post = await Post.findById(req.params.id)?.select('createdBy');
+            if(!post) res.status(404)
+            if (post.createdBy.toString() === req.user._id) await super.deleteById(req, res);
+            else {
+                res.status(401).send();
+            }
+        } catch (err) {
+            res.status(400).send(err.message);
+        }
     }
 
-    async like(req: AuthResquest, res: Response) {
+    async like(req: AuthRequest, res: Response) {
         const postId = req.params.id;
         try {
-            let requestedPost: IPost = await this.model.findById(postId);
-            if (requestedPost.usersWhoLiked.find(({ _id }) => _id === req.user._id)) {
+            let requestedPost: IPost = await this.model.findById(postId).select('usersWhoLiked');
+            if (requestedPost.usersWhoLiked.find(id => id === req.user._id)) {
                 throw ("failed, can't like an already liked post");
             } else {
                 const _id = req.user._id;
-                // req.body.owner = _id;
-                // requestedPost.usersWhoLiked.push(_id);
+                requestedPost.usersWhoLiked.push(_id);
                 const obj = await this.model.findByIdAndUpdate(postId,);
                 res.status(201).send(obj);
             }
@@ -64,17 +74,16 @@ class PostController extends BaseController<IPost>{
         }
     }
 
-    async unlike(req: AuthResquest, res: Response) {
+    async unlike(req: AuthRequest, res: Response) {
         const postId = req.params.id;
         try {
             let requestedPost: IPost = await this.model.findById(postId);
-            if (!requestedPost.usersWhoLiked.find(({ _id }) => _id === req.user._id)) {
+            if (!requestedPost.usersWhoLiked.find(id => id === req.user._id)) {
                 throw ("failed, can't unlike an already unliked post");
             } else {
                 const _id = req.user._id;
-                // req.body.owner = _id;
                 requestedPost.usersWhoLiked = requestedPost.usersWhoLiked.filter
-                    (({ _id }) => _id === req.user._id)
+                    (id => id === req.user._id)
                 const obj = await this.model.findByIdAndUpdate(postId,);
                 res.status(201).send(obj);
             }
@@ -84,7 +93,7 @@ class PostController extends BaseController<IPost>{
         }
     }
 
-    async comment(req: AuthResquest, res: Response) {
+    async comment(req: AuthRequest, res: Response) {
         const comment = req.body.comment;
         const postId = req.body.postId;
         try {
